Corrected force defrosting timer starting point
Force defrosting timer were starting in wrong place. It were starting in the place, when timeout happened for the waiting timer. 
This reseted the timer and pushed forced defrost forward, which would have leaded to really long operations. It also didn't keep the forced in sync with real defrozings.
Now only, when device is detected to perform defrozing, the timer is reseted.